Fix incorrect calculation of forward and up vectors in transform3d_to_fmod_vector()
This commit fixes a bug in the transform3d_to_fmod_vector() function where the calculation of the forward and up vectors was incorrect for certain cases. Specifically, for the VectorType::FORWARD case, the calculation was corrected to use the negative of the third column of the transform's basis matrix instead of the third row. Similarly, for the VectorType::UP case, the calculation was corrected to use the second column instead of the second row.